BB-2290: Handle on order save total discount and save to order

+++ src/OroB2B/Bundle/OrderBundle/Form/Type/EventListener/SubtotalSubscriber.php
@@ -121,6 +121,6 @@ class SubtotalSubscriber implements EventSubscriberInterface
      */
     protected function calculatePercent($subtotal, $discount)
     {
-        return (int) ($discount->getAmount() / $subtotal->getAmount() * 100);
+        return (float) ($discount->getAmount() / $subtotal->getAmount() * 100);
     }
 }
